fix: JangguStateWithTick update function bug fix
There was a bug that when user press hit key raise keyboard button, there was a two hit judgement(one at hit key, one at raise keyboard button). So I fixed it. When user hit and take stick off the janggu, state.궁채 is None, and self.궁채.face is not None. So that state is not hit event. So make toggle_keydown false.